requirements.txt
py-algorand-sdk==1.18.0
pyteal==0.18.1
beaker-pyteal==0.2.1